rename components to lowercase
