+++ app/contact/page.tsx
@@ -1,4 +1,4 @@
-import {Page} from '@/components/Page'
+import {Page} from '@/components/page'
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
 import {faHome, faPhone} from '@fortawesome/free-solid-svg-icons'
 import Link from 'next/link'
