+++ app/layout.tsx
@@ -2,7 +2,7 @@ import {Metadata} from 'next'
 import {Nunito} from 'next/font/google'
 import './globals.css'
 
-import { Page } from '@/components/Page'
+import { Page } from '@/components/page'
 
 const nextFont = Nunito({subsets: ['latin']})
